Add tentative support for fudged S-expressions.
This is a workaround to allow easier specification of symbols whose names conflict
with OCaml identifiers.

For example, the fudged S-expression (using `%sfxp` insteada of `sexp`):

    [%sfxp '\'',
       (let' ([a 7])
          (print a))]

will evaluate to

    `List [`Symbol "let"; `List [`List [`Symbol "a"; `Int 7]];
           `List[`Symbol "print"; `Symbol "a"]]

That is, the let' has been converted to let.

This is particularly applicable to the representation of Scheme expressions
because ' is not a valid character in Scheme identifiers, as it is in fact the
quote reader syntax.

This also helps with writing HTML forms using Sxmlm (form methods and input types).